+++ src/apps/reports/src/app/views/Analytics/Analytics.less
