+++ src/apps/reports/src/app/views/Analytics/components/ContentVelocity/ContentVelocity.js
