+++ src/apps/reports/src/app/views/Analytics/components/GoogleAuthOverlay/GaAuthenticate.js
