+++ src/apps/reports/src/app/views/Analytics/components/GoogleAuthOverlay/GoogleAuthOverlay.less
