+++ src/apps/reports/src/app/views/Analytics/components/InboundTraffic/InboundTraffic.less
