+++ src/apps/reports/src/app/views/Analytics/components/PageviewTraffic/PageviewTraffic.js
