+++ src/apps/reports/src/app/views/Analytics/components/RecentlyEdited/RecentlyEdited.less
