+++ src/apps/reports/src/app/views/Analytics/components/SocialTraffic/SocialTraffic.less
