+++ src/apps/reports/src/app/views/Analytics/components/SocialTraffic/index.js
