+++ src/apps/reports/src/app/views/Analytics/index.js
