+++ src/apps/reports/src/app/views/Analytics/shelldata.js
