+++ src/apps/reports/src/app/views/AuditTrail/components/controls/styles.less
