+++ src/apps/reports/src/app/views/AuditTrail/components/log/styles.less
