+++ src/apps/reports/src/app/views/AuditTrail/components/pagination/styles.less
