+++ src/apps/reports/src/app/views/AuditTrail/styles.less
