+++ src/apps/reports/src/app/views/Metrics/Metrics.less
