+++ src/apps/reports/src/store/logsInView.js
